services/pr-orchestrator/requirements.txt
# Web Framework
fastapi
uvicorn

# Google Cloud Libraries
google-cloud-pubsub
google-cloud-firestore